fix: Fixed another warning

+++ arch/eventloop_posix/eventloop_posix.c
@@ -783,7 +783,7 @@ UA_EventLoopPOSIX_pollFDs(UA_EventLoopPOSIX *el, UA_DateTime listenTimeout) {
     };
 
     UA_UNLOCK(&el->elMutex);
-    int selectStatus = UA_select(highestfd + 1, &readset, &writeset, &errset, &tmptv);
+    int selectStatus = UA_select((int)highestfd + 1, &readset, &writeset, &errset, &tmptv);
     UA_LOCK(&el->elMutex);
     if(selectStatus < 0) {
         /* We will retry, only log the error */
